validator_api/requirements.txt
django
djangorestframework
requests
django-health-check